+++ data/src/main/java/com/gridnine/testing/Segment.java
@@ -1,0 +1,37 @@
+package com.gridnine.testing;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.Objects;
+
+/**
+ * Bean that represents a flight segment.
+ */
+class Segment {
+
+    private final LocalDateTime departureDate;
+
+    private final LocalDateTime arrivalDate;
+
+    Segment(final LocalDateTime dep, final LocalDateTime arr) {
+        departureDate = Objects.requireNonNull(dep);
+        arrivalDate = Objects.requireNonNull(arr);
+    }
+
+    LocalDateTime getDepartureDate() {
+        return departureDate;
+    }
+
+    LocalDateTime getArrivalDate() {
+        return arrivalDate;
+    }
+
+    @Override
+    public String toString() {
+        DateTimeFormatter fmt =
+                DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm");
+        return '[' + departureDate.format(fmt) + '|' + arrivalDate.format(fmt)
+                + ']';
+    }
+
+}
